Final comit of Sprint 7.
Some things to have in count: The play button disappears after switching of fragment.
The colection fragment is yet to be finished, and the clans fragment yet to be done.

Now the game has playability, altough is very simple.

We are open to advices to improve the game this week, to have a better product on friday